src/ai_models/model.py
"""
AI Model Training Module
This module demonstrates various AI framework imports and configurations.
"""

import torch
import tensorflow as tf
from transformers import AutoModelForCausalLM, AutoTokenizer
from langchain import OpenAI, LLMChain
import openai
from anthropic import Anthropic
import pinecone
from sentence_transformers import SentenceTransformer

# Configuration (mock keys)
OPENAI_KEY = "sk-abcd1234efgh5678ijkl9012mnop3456qrst7890uvwx1234yzab"
ANTHROPIC_KEY = "sk-ant-abcd1234efgh5678ijkl9012mnop3456qrst7890"

class AIModel:
    def __init__(self):
        self.device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
        self.model = None
        self.tokenizer = None
        
    def load_model(self):
        """Load pretrained model and tokenizer"""
        self.model = AutoModelForCausalLM.from_pretrained("gpt2")
        self.tokenizer = AutoTokenizer.from_pretrained("gpt2")
        
    def setup_vector_db(self):
        """Initialize vector database connection"""
        pinecone.init(api_key="pinecone-abcdefghijklmnopqrstuvwxyz")
        
    def generate_embeddings(self, text):
        """Generate embeddings using sentence transformers"""
        model = SentenceTransformer('all-MiniLM-L6-v2')
        return model.encode(text)

if __name__ == "__main__":
    model = AIModel()
    model.load_model() 