requirements.txt
tensorflow>=2.13.0
torch>=2.0.0
transformers>=4.30.0
langchain>=0.0.200
openai>=0.27.0
anthropic>=0.3.0
pinecone-client>=2.2.2
sentence-transformers>=2.2.2
huggingface-hub>=0.16.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
jupyter>=1.0.0
python-dotenv>=1.0.0 